+++ lib/pmdtester/builders/pmd_report_builder.rb
@@ -2,6 +2,8 @@ require 'fileutils'
 require_relative '../cmd'
 include PmdTester
 module PmdTester
+  # Building pmd xml reports according to a list of standard
+  # projects and branch of pmd source code
   class PmdReportBuilder
     def initialize(branch_config, projects, local_git_repo, pmd_branch_name)
       @branch_config = branch_config
@@ -12,15 +14,15 @@ module PmdTester
     end
 
     def create_repositories_dir
-      @repositories_dir = "#@pwd/target/repositories"
+      @repositories_dir = "#{@pwd}/target/repositories"
       FileUtils.mkdir_p(@repositories_dir) unless File.directory?(@repositories_dir)
     end
 
     def execute_reset_cmd(type, tag)
       case type
-      when "git"
+      when 'git'
         reset_cmd = "git reset --hard #{tag}"
-      when "hg"
+      when 'hg'
         reset_cmd = "hg up #{tag}"
       else
         raise Exception, "Unknown #{type} repository"
@@ -35,38 +37,36 @@ module PmdTester
       create_repositories_dir
 
       @projects.each do |project|
-        path = "#@repositories_dir/#{project.name}"
+        path = "#{@repositories_dir}/#{project.name}"
         clone_cmd = "#{project.type} clone #{project.connection} #{path}"
-        if File::exist?(path)
+        if File.exist?(path)
           puts "Skipping clone, project path #{path} already exists"
         else
           Cmd.execute(clone_cmd)
         end
         project.local_path = path
 
-        unless project.tag.nil?
-          Dir.chdir(path) do
-            execute_reset_cmd(project.type, project.tag)
-          end
+        next if project.tag.nil?
+        Dir.chdir(path) do
+          execute_reset_cmd(project.type, project.tag)
         end
       end
     end
 
     def get_pmd_binary_file
       Dir.chdir(@local_git_repo) do
-
-        checkout_cmd = "git checkout #@pmd_branch_name"
+        checkout_cmd = "git checkout #{@pmd_branch_name}"
         Cmd.execute(checkout_cmd)
 
-        package_cmd = './mvnw clean package -Dpmd.skip=true -Dmaven.test.skip=true' +
-            ' -Dmaven.checkstyle.skip=true -Dmaven.javadoc.skip=true'
+        package_cmd = './mvnw clean package -Dpmd.skip=true -Dmaven.test.skip=true' \
+                      ' -Dmaven.checkstyle.skip=true -Dmaven.javadoc.skip=true'
         Cmd.execute(package_cmd)
 
-        version_cmd = "./mvnw -q -Dexec.executable=\"echo\" -Dexec.args='${project.version}' " +
-            "--non-recursive org.codehaus.mojo:exec-maven-plugin:1.5.0:exec"
+        version_cmd = "./mvnw -q -Dexec.executable=\"echo\" -Dexec.args='${project.version}' " \
+                      '--non-recursive org.codehaus.mojo:exec-maven-plugin:1.5.0:exec'
         @pmd_version = Cmd.execute(version_cmd)
 
-        target_dir = "#@pwd/target"
+        target_dir = "#{@pwd}/target"
         unzip_cmd = "unzip -qo pmd-dist/target/pmd-bin-#{@pmd_version}.zip -d #{target_dir}"
         Cmd.execute(unzip_cmd)
       end
@@ -74,7 +74,8 @@ module PmdTester
 
     def generate_pmd_report(src_root_dir, report_file)
       run_path = "target/pmd-bin-#{@pmd_version}/bin/run.sh"
-      pmd_cmd = "#{run_path} pmd -d #{src_root_dir} -f xml -R #@branch_config -r #{report_file} -failOnViolation false"
+      pmd_cmd = "#{run_path} pmd -d #{src_root_dir} -f xml -R #{@branch_config} " \
+                "-r #{report_file} -failOnViolation false"
       Cmd.execute(pmd_cmd)
     end
 
@@ -85,7 +86,7 @@ module PmdTester
 
       pmd_branch_name = @pmd_branch_name.delete('/')
       branch_file = "target/reports/#{pmd_branch_name}"
-      FileUtils::mkdir_p(branch_file) unless File.directory?(branch_file)
+      FileUtils.mkdir_p(branch_file) unless File.directory?(branch_file)
 
       @projects.each do |project|
         project_report_file = "#{branch_file}/#{project.name}.xml"